+++ app.config.js
@@ -20,7 +20,7 @@ class AppConfig{
 
     get BOT(){ return require('./src/engine/bot')};
     get EVENTS(){ return require('./src/engine/events')};
-    get COMMANDS(){ return require('./src/engine/commands')};
+    get COMMANDS(){ return require('./src/engine/commands/commands')};
     get PERMISSIONS(){ return require('./src/engine/permissions')};
 }
